feat: add BankAccount to RegistrationCustomer and ReasonFailure in TransactionCompletedEvent

+++ UserService/UserService.Domain/Entities/BankAccount.cs
@@ -4,7 +4,7 @@
     {
         public Guid Id { get; set; }
         public Guid CustomerId { get; set; }
-        public int MoneyAmount { get; set; }
+        public long MoneyAmount { get; set; }
         public string? AccountType { get; set; }
 
     }
